Выполнили 4 домашку
